feat!: Better error handling
Allows the library users handling async errors by promises and subscribing to `error` events.

BREAKING CHANGE: The applications have to subscribe to the `error` event, otherwise the process is terminated in case of an error. Also all calls that return Promise have to be chained with `catch` or await/try/catch.

Closes Error: error:06065064:digital envelope routines:EVP_DecryptFinal_ex:bad decrypt inwaar/gree-hvac-client#13